pyproject.toml
[tool.poetry]
name = "chat-langchain"
version = "0.1.1"
description = ""
authors = ["SN <6432132+samnoyes@users.noreply.github.com>"]
readme = "README.md"
packages = [{ include = "backend" }]

[tool.poetry.dependencies]
python = "^3.12"
langchain = ">=0.3.0,<0.4.0"
langsmith = { version = "0.2.10", extras = ["langsmith-pyo3", "compression"] }
langchain-cohere = ">=0.3.0,<0.4.0"
langchain-google-genai = ">=2.0.0,<3.0.0"
langchain-core = ">=0.3.10,<0.4.0"
langchain-community = ">=0.3.0,<0.4.0"
langchain-openai = ">=0.2.0,<0.3.0"
langchain-anthropic = ">=0.2.0,<0.3.0"
langchain-fireworks = ">=0.2.0,<0.3.0"
langchain-groq = ">=0.2.0,<0.3.0"
langchain-weaviate = ">=0.0.3,<0.1.0"
langgraph = ">=0.2.32,<0.3.0"
beautifulsoup4 = "^4.12.2"
weaviate-client = "^4.0.0"
lxml = "^4.9.3"
voyageai = "^0.1.4"
pillow = "^10.2.0"
# do not remove this, it's used for indexing
psycopg2-binary = "^2.9.9"
uvicorn = "^0.34.0"
fastapi = "^0.115.6"
sse-starlette = "^2.2.1"
rich = "^13.9.4"
motor = "^3.6.1"
pymongo = ">=4.9,<4.10"
azure-search-documents = "11.4.0"
azure-identity = "1.15.0"
azure-core = "1.30.0"
pypdf = "^5.3.0"
python-docx = "^1.1.0"
python-pptx = "^1.0.1"
dotenv = "^0.9.9"
requests = "^2.31.0"
easyocr = "^1.7.1"
PyMuPDF = "^1.23.8"
pdfplumber = "^0.10.3"
reportlab = "^4.0.8"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.0"
ruff = "^0.2.2"
pandas = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
